Not allow delete subnet gateway port
There is no check about delete gateway port. User can delete it through
API directly.

So this patch add the check like ml2.

Closes-Bug: #1531681

+++ dragonflow/tests/fullstack/test_objects.py
@@ -33,6 +33,8 @@ class RouterTestWrapper(object):
         ports = self.neutron.list_ports(device_id=self.router_id)
         ports = ports['ports']
         for port in ports:
+            if port['device_owner'] == 'network:router_interface':
+                continue
             self.neutron.delete_port(port['id'])
         self.neutron.delete_router(self.router_id)
         self.deleted = True
